+++ client-react/src/pages/app/deployment-center/DeploymentCenterFtps.tsx
@@ -1,5 +1,4 @@
 import React, { useContext, useState } from 'react';
-import { Field } from 'formik';
 import {
   DeploymentCenterFtpsProps,
   DeploymentCenterFieldProps,
@@ -10,7 +9,6 @@ import { MessageBarType, ActionButton, ProgressIndicator } from 'office-ui-fabri
 import { useTranslation } from 'react-i18next';
 import { deploymentCenterContent, additionalTextFieldControl, deploymentCenterInfoBannerDiv } from './DeploymentCenter.styles';
 import TextFieldNoFormik from '../../../components/form-controls/TextFieldNoFormik';
-import TextField from '../../../components/form-controls/TextField';
 import CustomBanner from '../../../components/CustomBanner/CustomBanner';
 import { DeploymentCenterContext } from './DeploymentCenterContext';
 import CustomFocusTrapCallout from '../../../components/CustomCallout/CustomFocusTrapCallout';
@@ -18,6 +16,7 @@ import { Links } from '../../../utils/FwLinks';
 import { DeploymentCenterPublishingContext } from './DeploymentCenterPublishingContext';
 import { ScmType } from '../../../models/site/config';
 import { getGitCloneUri } from './utility/DeploymentCenterUtility';
+import DeploymentCenterPublishingUser from './DeploymentCenterPublishingUser';
 
 type PasswordFieldType = 'password' | undefined;
 
@@ -28,30 +27,15 @@ const DeploymentCenterFtps: React.FC<
   const deploymentCenterContext = useContext(DeploymentCenterContext);
   const deploymentCenterPublishingContext = useContext(DeploymentCenterPublishingContext);
 
-  const { isLoading } = props;
-  const { publishingProfile, publishingUser, resetApplicationPassword } = deploymentCenterPublishingContext;
+  const { isLoading, formProps } = props;
+  const { publishingProfile, resetApplicationPassword } = deploymentCenterPublishingContext;
 
   const [applicationPasswordType, setApplicationPasswordType] = useState<PasswordFieldType>('password');
-  const [providerPasswordType, setProviderPasswordType] = useState<PasswordFieldType>('password');
-  const [providerConfirmPasswordType, setProviderConfirmPasswordType] = useState<PasswordFieldType>('password');
+
   const [isResetCalloutHidden, setIsResetCalloutHidden] = useState<boolean>(true);
   const [showBlockedBanner, setShowBlockedBanner] = useState(true);
 
   const ftpsEndpoint = publishingProfile && publishingProfile.publishUrl.toLocaleLowerCase().replace('ftp:/', 'ftps:/');
-  const webProviderUsername = publishingUser && publishingUser.properties.publishingUserName;
-
-  const siteDescriptor = deploymentCenterContext && deploymentCenterContext.siteDescriptor;
-  const sampleAppNameDomain =
-    siteDescriptor && siteDescriptor.slot
-      ? `${siteDescriptor.site}-${siteDescriptor.slot}`
-      : siteDescriptor && siteDescriptor.site
-      ? siteDescriptor.site
-      : '';
-
-  const sampleWebProviderDomainUsername = webProviderUsername
-    ? `${sampleAppNameDomain}\\${webProviderUsername}`
-    : `${sampleAppNameDomain}\\${t('deploymentCenterFtpsUserScopeSampleUsername')}`;
-
   const isScmLocalGit = deploymentCenterContext.siteConfig && deploymentCenterContext.siteConfig.properties.scmType === ScmType.LocalGit;
   const gitCloneUri = getGitCloneUri(deploymentCenterPublishingContext);
 
@@ -66,14 +50,6 @@ const DeploymentCenterFtps: React.FC<
 
   const toggleShowApplicationPassword = () => {
     setApplicationPasswordType(!applicationPasswordType ? 'password' : undefined);
-  };
-
-  const toggleShowProviderPassword = () => {
-    setProviderPasswordType(!providerPasswordType ? 'password' : undefined);
-  };
-
-  const toggleShowProviderConfirmPassword = () => {
-    setProviderConfirmPasswordType(!providerConfirmPasswordType ? 'password' : undefined);
   };
 
   const getProgressIndicator = () => {
@@ -186,60 +162,7 @@ const DeploymentCenterFtps: React.FC<
           defaultButtonFunction={toggleResetCalloutVisibility}
         />
 
-        <h3>{t('deploymentCenterFtpsUserScopeTitle')}</h3>
-        <p>{t('deploymentCenterFtpsUserScopeDescription').format(sampleWebProviderDomainUsername)}</p>
-
-        <Field
-          id="deployment-center-ftps-provider-username"
-          name="publishingUsername"
-          component={TextField}
-          widthOverride="60%"
-          label={t('deploymentCenterFtpsUsernameLabel')}
-        />
-
-        <Field
-          id="deployment-center-ftps-provider-password"
-          name="publishingPassword"
-          component={TextField}
-          widthOverride="60%"
-          label={t('deploymentCenterFtpsPasswordLabel')}
-          type={providerPasswordType}
-          additionalControls={[
-            <ActionButton
-              id="deployment-center-ftps-provider-password-visibility-toggle"
-              key="deployment-center-ftps-provider-password-visibility-toggle"
-              className={additionalTextFieldControl}
-              ariaLabel={providerPasswordType === 'password' ? t('showProviderPasswordAriaLabel') : t('hideProviderPasswordAriaLabel')}
-              onClick={toggleShowProviderPassword}
-              iconProps={{ iconName: providerPasswordType === 'password' ? 'RedEye' : 'Hide' }}>
-              {providerPasswordType === 'password' ? t('show') : t('hide')}
-            </ActionButton>,
-          ]}
-        />
-
-        <Field
-          id="deployment-center-ftps-provider-confirm-password"
-          name="publishingConfirmPassword"
-          component={TextField}
-          widthOverride="60%"
-          label={t('deploymentCenterFtpsConfirmPasswordLabel')}
-          type={providerConfirmPasswordType}
-          additionalControls={[
-            <ActionButton
-              id="deployment-center-ftps-provider-confirm-password-visibility-toggle"
-              key="deployment-center-ftps-provider-confirm-password-visibility-toggle"
-              className={additionalTextFieldControl}
-              ariaLabel={
-                providerConfirmPasswordType === 'password'
-                  ? t('showProviderConfirmPasswordAriaLabel')
-                  : t('hideProviderConfirmPasswordAriaLabel')
-              }
-              onClick={toggleShowProviderConfirmPassword}
-              iconProps={{ iconName: providerConfirmPasswordType === 'password' ? 'RedEye' : 'Hide' }}>
-              {providerConfirmPasswordType === 'password' ? t('show') : t('hide')}
-            </ActionButton>,
-          ]}
-        />
+        <DeploymentCenterPublishingUser formProps={formProps} isLoading={isLoading} />
       </div>
     );
   };
