+++ client/src/app/shared/models/portal-resources.ts
@@ -1991,6 +1991,9 @@ export class PortalResources {
   public static deploymentCenterFtpsPasswordLabel = 'deploymentCenterFtpsPasswordLabel';
   public static deploymentCenterFtpsConfirmPasswordLabel = 'deploymentCenterFtpsConfirmPasswordLabel';
   public static deploymentCenterFtpsWritePermissionRequired = 'deploymentCenterFtpsWritePermissionRequired';
+  public static deploymentCenterPublishingUserLoadingMessage = 'deploymentCenterPublishingUserLoadingMessage';
+  public static deploymentCenterPublishingUserLoadingAriaLabel = 'deploymentCenterPublishingUserLoadingAriaLabel';
+  public static publishingUserFetchFailedMessage = 'publishingUserFetchFailedMessage';
   public static appKeys_hostKeys_description = 'appKeys_hostKeys_description';
   public static appKeys_systemKeys_description = 'appKeys_systemKeys_description';
   public static functionKeys_description = 'functionKeys_description';
